+++ src/main/java/dev/abarmin/aml/exception/ExceptionAdvice.java
@@ -7,6 +7,12 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @Slf4j
 @ControllerAdvice
 public class ExceptionAdvice {
+
+  @ExceptionHandler(OidcRegistrationNotCompletedException.class)
+  public String oidcRegistrationNotCompleted() {
+    return "redirect:/register/oidc";
+  }
+
   @ExceptionHandler(Exception.class)
   public String serverError(Exception e) {
     log.warn("Server error", e);