+++ src/main/java/dev/abarmin/aml/registration/RegistrationFormValidator.java
@@ -3,6 +3,7 @@ package dev.abarmin.aml.registration;
 import dev.abarmin.aml.registration.repository.ProfileRepository;
 import dev.abarmin.aml.registration.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
@@ -27,6 +28,15 @@ public class RegistrationFormValidator implements Validator {
       .ifPresent(profile -> errors.rejectValue("link", "link.exists"));
     if (!form.isAcceptTerms()) {
       errors.rejectValue("acceptTerms", "acceptTerms.required");
+    }
+    if (form.isPasswordRequired()) {
+      if (StringUtils.isEmpty(form.getPassword())) {
+        errors.rejectValue("password", "password.required");
+      } else if (form.getPassword().length() < 3) {
+        errors.rejectValue("password", "password.tooShort");
+      } else if (form.getPassword().length() > 10) {
+        errors.rejectValue("password", "password.tooLong");
+      }
     }
   }
 }