+++ src/main/java/dev/abarmin/aml/registration/domain/AccountType.java
@@ -1,5 +1,6 @@
 package dev.abarmin.aml.registration.domain;
 
 public enum AccountType {
-  USERNAME_PASSWORD
+  USERNAME_PASSWORD,
+  OIDC
 }