+++ src/main/resources/db/migration/V11__create_column_profile_qr.sql
@@ -1,3 +1,0 @@
-ALTER TABLE user_profiles
-ADD COLUMN IF NOT EXISTS profile_qr VARCHAR(255) NULL
-AFTER profile_link;